Refactoring to add tests
- dumping express-routescan, string-format dependencies
- adding dev dependencies for testing
- adding testing scripts
- refactoring a few files so they play well with tests
- gtfs-sequelize update to 0.2.3
