+++ server/routes/index.js
@@ -1,0 +1,9 @@
+var fs = require('fs');
+
+module.exports = function(app, config){
+  fs.readdirSync(__dirname).forEach(function(file) {
+    if (file == "index.js") return;
+    var name = file.substr(0, file.indexOf('.'));
+    require('./' + name)(app, config);
+  });
+}
